rename and restyle, remove tiny files used in one place
