+++ static/app/components/events/interfaces/threads/threadSelector/styles.tsx
@@ -2,16 +2,15 @@ import styled from '@emotion/styled';
 
 import {space} from 'sentry/styles/space';
 
-const Grid = styled('div')<{hasThreadStates: boolean}>`
+export const ThreadSelectorGrid = styled('div')<{hasThreadStates: boolean}>`
   font-size: ${p => p.theme.fontSizeSmall};
   display: grid;
-  gap: ${space(1)};
+  gap: ${space(0.5)};
   align-items: center;
   grid-template-columns: 16px 0.5fr repeat(${p => (p.hasThreadStates ? '2' : '1')}, 1fr) 1fr;
+  min-height: 18px;
 `;
 
-const GridCell = styled('div')`
+export const ThreadSelectorGridCell = styled('div')`
   ${p => p.theme.overflowEllipsis};
 `;
-
-export {Grid, GridCell};
