--- static/app/components/events/interfaces/threads/threadSelector/trimFilename.tsx
@@ -1,6 +1,0 @@
-function trimFilename(filename: string) {
-  const pieces = filename.split(/\//g);
-  return pieces[pieces.length - 1];
-}
-
-export default trimFilename;
